requirements-multitenant.txt
# StormOps Multi-Tenant Platform Requirements

# Core
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0

# UI
streamlit>=1.28.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Observability
prometheus-client>=0.18.0

# Future: Streaming
# kafka-python>=2.0.2
# apache-flink>=1.17.0

# Future: ML
# scikit-learn>=1.3.0
# xgboost>=2.0.0

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0
